tagcache compress uniqbuf 2 16-bit indices occupy a single 32 bit slot
a lot of space is wasted when file indices less than 65535
entries should be more than 1 byte apart so use the LSB as a flag
when indices are > 65535 set flag to 0 and proceed as before

explicitly mark uniqbuf as 32bit

+++ apps/tagcache.h
@@ -188,7 +188,7 @@ struct tagcache_search {
     int entry_count;
     bool valid;
     bool initialized;
-    unsigned long *unique_list;
+    uint32_t *unique_list;
     int unique_list_capacity;
     int unique_list_count;
